Latest fixes
